HRMS_backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pydantic[email]